@@ -38,7 +38,10 @@
   
   <dependencies>
     
-    
+    <dependency>
+      <groupId>io.swagger.core.v3</groupId>
+      <artifactId>swagger-annotations-jakarta</artifactId>
+    </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-core</artifactId>
@@ -93,10 +96,6 @@
       <artifactId>spring-webmvc</artifactId>
     </dependency>
     <dependency>
-      <groupId>javax.annotation</groupId>
-      <artifactId>javax.annotation-api</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-aop</artifactId>
       <exclusions>
@@ -128,15 +127,13 @@
     <dependency>
       <groupId>cn.hutool</groupId>
       <artifactId>hutool-all</artifactId>
-      <version>5.7.20</version>
-      <!--      <scope>provided</scope>-->
     </dependency>
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>servlet-api</artifactId>
-      <version>2.5</version>
-      <scope>provided</scope>
-    </dependency>
+<!--    <dependency>-->
+<!--      <groupId>javax.servlet</groupId>-->
+<!--      <artifactId>servlet-api</artifactId>-->
+<!--      <version>2.5</version>-->
+<!--      <scope>provided</scope>-->
+<!--    </dependency>-->
     
     
     <dependency>
@@ -160,17 +157,15 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.8.1</version>
         <configuration>
           <source>17</source>
           <target>17</target>
-          <encoding>utf-8</encoding>
+          <encoding>UTF-8</encoding>
         </configuration>
       </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
-        <version>2.18.1</version>
         <configuration>
           <skipTests>true</skipTests>
         </configuration>
@@ -214,7 +209,6 @@
           <plugin>
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-source-plugin</artifactId>
-            <version>3.2.1</version>
             <executions>
               <execution>
                 <phase>package</phase>
@@ -230,7 +224,7 @@
             <artifactId>maven-javadoc-plugin</artifactId>
             <version>3.2.0</version>
             <configuration>
-              <source>8</source>
+              <source>17</source>
               <show>private</show>
               <nohelp>true</nohelp>
               <charset>UTF-8</charset>
@@ -269,10 +263,9 @@
           <plugin>
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-compiler-plugin</artifactId>
-            <version>3.8.1</version>
             <configuration>
-              <source>1.8</source>
-              <target>1.8</target>
+              <source>17</source>
+              <target>17</target>
               <fork>true</fork>
               <verbose>true</verbose>
               <encoding>UTF-8</encoding>
