@@ -4,7 +4,7 @@ import com.fairychar.bag.function.Action;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
