@@ -14,22 +14,22 @@ import java.time.format.DateTimeFormatter;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class Consts {
 
-    public static String EMPTY_STR = "";
-    public static long KB_PER_B = 1024;
-    public static long MB_PER_B = 1024 * 1024;
-    public static long GB_PER_B = 1024 * 1024 * 1024;
-    public static long TB_PER_B = 1024 * 1024 * 1024 * 1024;
-    public static long PB_PER_B = 1024 * 1024 * 1024 * 1024 * 1024;
+    public static final String EMPTY_STR = "";
+    public static final long KB_PER_B = 1024;
+    public static final long MB_PER_B = 1024 * 1024;
+    public static final long GB_PER_B = 1024 * 1024 * 1024;
+    public static final long TB_PER_B = 1024 * 1024 * 1024 * 1024;
+    public static final long PB_PER_B = 1024 * 1024 * 1024 * 1024 * 1024;
 
-    public static String NONE = "none";
+    public static final String NONE = "none";
 
-    public static String SIMPLE_DATETIME_FORMAT = "yyyy-MM-dd HH:mm:ss";
+    public static final String SIMPLE_DATETIME_FORMAT = "yyyy-MM-dd HH:mm:ss";
 
-    public static DateTimeFormatter SIMPLE_DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern(SIMPLE_DATETIME_FORMAT);
+    public static final DateTimeFormatter SIMPLE_DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern(SIMPLE_DATETIME_FORMAT);
 
-    public static String SIMPLE_DATE_FORMAT = "yyyy-MM-dd";
+    public static final String SIMPLE_DATE_FORMAT = "yyyy-MM-dd";
 
-    public static DateTimeFormatter SIMPLE_DATE_FORMATTER = DateTimeFormatter.ofPattern(SIMPLE_DATE_FORMAT);
+    public static final DateTimeFormatter SIMPLE_DATE_FORMATTER = DateTimeFormatter.ofPattern(SIMPLE_DATE_FORMAT);
 
 
     public static final class OAuth2 {
@@ -41,8 +41,8 @@ public final class Consts {
     }
 
     public static final class Regex {
-        public static String IP = "^(\\d{1,3}\\.){3}\\d{1,3}$";
-        public static String URL = "^(http|https):\\/\\/[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}([\\/a-zA-Z0-9#.-]*)*$";
+        public static final String IP = "^(\\d{1,3}\\.){3}\\d{1,3}$";
+        public static final String URL = "^(http|https):\\/\\/[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}([\\/a-zA-Z0-9#.-]*)*$";
     }
 }
 /*
