@@ -1,10 +1,10 @@
 package com.fairychar.bag.domain.validator.rest;
 
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
