@@ -1,6 +1,6 @@
 package com.fairychar.bag.pojo.query.body;
 
-import io.swagger.annotations.ApiModel;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -17,7 +17,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-@ApiModel(value = "StringList", description = "String类型Json请求集合")
+@Schema(name = "StringList", description = "String类型Json请求集合")
 public final class StringListQuery implements Serializable {
     private List<String> list;
 }
